requirements.txt
Django==1.7
Pillow==2.5.3
beautifulsoup4==4.1.3
dj-database-url==0.3.0
django-blog-zinnia==0.14.2
django-contrib-comments==1.5
django-mptt==0.6.1
django-tagging==0.3.2
django-xmlrpc==0.1.5
gunicorn==19.1.1
psycopg2==2.5.4
pyparsing==2.0.2
pytz==2014.7
requests==2.4.1
static3==0.5.1
unittest2py3k==0.5.1


# To get around OSX compile errors when installing psycopg2, run this first.
#export ARCHFLAGS="-Wno-error=unused-command-line-argument-hard-error-in-future"
